Move parameter display helpers to shared util module
Move getResourceStateDescription/getParameterValueDescription and the mask/truncation constants out of the AspireAppHostTreeProvider view module into a new extension/src/utils/resourceDisplay.ts, so AspireCodeLensProvider no longer depends on the view layer. Pure move; no behavior change.
